Bug - VMALLOC
Fix a bug on vmalloc -> loop to set block of memory of growing
to fast due to a mistake on incrementation.
Also fixed an out of range error in vmalloc.

pmm stack fixed as a stack rather than a queue, at this point,
get_page and free_page are taking the hand and working as the
previous implementation, but with complexity O(1).

remain to do -
	- managing low memory
	- get_pages and free_pages
	- think about defragmentation in some kind of way ?